Move checkThat to index
