+++ packages/wallet-specs/src/index.ts
@@ -136,3 +136,16 @@ export const FINAL = 'final' as 'final';
 export function outcomesEqual(left: Outcome, right: Outcome): boolean {
   return JSON.stringify(left) === JSON.stringify(right);
 }
+
+const throwError = (fn: (t1: any) => boolean, t) => {
+  throw new Error(`not valid, ${fn.name} failed on ${t}`);
+};
+type TypeGuard<T, S> = (t1: T | S) => t1 is T;
+export function checkThat<T, S>(t: T | S, isTypeT: TypeGuard<T, S>): T {
+  if (!isTypeT(t)) {
+    throwError(isTypeT, t);
+    // Typescrypt doesn't know that throwError throws an error.
+    throw 'Unreachable';
+  }
+  return t;
+}
