+++ packages/wallet-specs/src/store.ts
@@ -289,16 +289,3 @@ export interface Deposit {
 }
 
 export type StoreEvent = ChannelUpdated | Deposit;
-
-const throwError = (fn: (t1: any) => boolean, t) => {
-  throw new Error(`not valid, ${fn.name} failed on ${t}`);
-};
-type TypeGuard<T, S> = (t1: T | S) => t1 is T;
-export function checkThat<T, S>(t: T | S, isTypeT: TypeGuard<T, S>): T {
-  if (!isTypeT(t)) {
-    throwError(isTypeT, t);
-    // Typescrypt doesn't know that throwError throws an error.
-    throw 'Unreachable';
-  }
-  return t;
-}
